INS-3011 use thread safe t in concurency test
