+++ functest/utils_test.go
@@ -29,7 +29,6 @@ import (
 	"regexp"
 	"runtime"
 	"strings"
-	"sync"
 	"testing"
 
 	"github.com/insolar/insolar/api"
@@ -510,97 +509,3 @@ func callMethodNoChecks(t testing.TB, objectRef *insolar.Reference, method strin
 
 	return callRes
 }
-
-type SyncT struct {
-	*testing.T
-
-	mu sync.Mutex
-}
-
-var _ testing.TB = (*SyncT)(nil)
-
-func (t SyncT) Error(args ...interface{}) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	t.T.Error(args...)
-}
-func (t SyncT) Errorf(format string, args ...interface{}) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	t.T.Errorf(format, args...)
-}
-func (t SyncT) Fail() {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	t.T.Fail()
-}
-func (t SyncT) FailNow() {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	t.T.FailNow()
-}
-func (t SyncT) Failed() bool {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	return t.T.Failed()
-}
-func (t SyncT) Fatal(args ...interface{}) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	t.T.Fatal(args...)
-}
-func (t SyncT) Fatalf(format string, args ...interface{}) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	t.T.Fatalf(format, args...)
-}
-func (t SyncT) Log(args ...interface{}) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	t.T.Log(args...)
-}
-func (t SyncT) Logf(format string, args ...interface{}) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	t.T.Logf(format, args...)
-}
-func (t SyncT) Name() string {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	return t.T.Name()
-}
-func (t SyncT) Skip(args ...interface{}) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	t.T.Skip(args...)
-}
-func (t SyncT) SkipNow() {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	t.T.SkipNow()
-}
-func (t SyncT) Skipf(format string, args ...interface{}) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	t.T.Skipf(format, args...)
-}
-func (t SyncT) Skipped() bool {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	return t.T.Skipped()
-}
-func (t SyncT) Helper() {}
